+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.17.0] - 2026-06-28
+
 ### Added
 
 - Gekennzeichneten Zeitraum bearbeiten: `PUT /api/periods/{period_id}` ändert einen